requirements.txt
gradio
Pillow
matplotlib
numpy
